pyproject.toml
[project]
name = "cinema-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.8.1",
    "django==5.1.4",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.3.1",
    "mkdocs>=1.6.1",
    "pyjwt==2.10.1",
    "sqlparse==0.5.3",
    "typing-extensions==4.12.2",
]
